pyproject.toml
[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "life-as-code"
version = "0.1.0"
description = "Personal health tracking and analysis system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Life-as-Code Core Dependencies
    "pandas>=2.3.1",
    "garminconnect>=0.2.38",
    "requests>=2.32.4",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    # Database Dependencies
    "sqlalchemy>=2.0.42",
    "psycopg2-binary>=2.9.10",
    "flask-migrate>=4.0.5",
    "alembic>=1.13.0",
    # Authentication & Security Dependencies
    "Flask>=3.1.1",
    "Flask-Login>=0.6.3",
    "cryptography>=45.0.5",
    "bcrypt>=4.0.0",
    "flask-wtf>=1.2.1",
    "flask-limiter>=3.8.0",
    "redis>=5.0.0",
    # Data Validation & Settings
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    # Structured Logging
    "structlog>=24.4.0",
    # Retry logic
    "tenacity>=9.0.0",
    # Production WSGI Server
    "gunicorn>=23.0.0",
    # Transitive dependency security fix (GHSA-w853-jp5j-5j7f)
    "filelock>=3.20.1",
]

[project.optional-dependencies]
ml = [
    "chronos-forecasting>=2.1.0",
    "torch>=2.2,<3",
    "transformers>=4.40",
    "scikit-learn>=1.5",
    "scipy>=1.12",
    "numpy>=1.26",
]
agent = [
    "anthropic>=0.40",
]
bot = [
    "python-telegram-bot[job-queue]>=21.0",
]
all = [
    "life-as-code[ml,agent,bot]",
]
dev = [
    "pre-commit>=3.5.0",
    "black>=24.1.0",
    "ruff>=0.1.15",
    "mypy>=1.8.0",
    # Type Checking Stubs
    "types-PyYAML>=6.0.12.20250516",
    "types-requests>=2.32.4.20250611",
]

[tool.setuptools.packages.find]
where = ["src"]
